+++ src/BKS.auto.php
@@ -8,8 +8,8 @@
  * @author    Puneet Gopinath (PuneetGopinath) <baalkrshna@gmail.com>
  * @copyright 2021 The BK Sanitizers Team
  * @license   http://www.opensource.org/licenses/MIT MIT
- * @see       https://github.com/PuneetGopinath/Sanitizers BKS on GitHub
- * @see       https://packagist.org/packages/sanitizers/sanitizers BKS on Packagist
+ * @see       https://github.com/PuneetGopinath/Sanitizers BKS - GitHub
+ * @see       https://packagist.org/packages/sanitizers/sanitizers BKS - Packagist
  */
 
 $dir = dirname(__FILE__);