+++ src/Sanitizer.php
@@ -8,8 +8,8 @@
  * @author    Puneet Gopinath (PuneetGopinath) <baalkrshna@gmail.com>
  * @copyright 2021 The BK Sanitizers Team
  * @license   http://www.opensource.org/licenses/MIT MIT
- * @see       https://github.com/PuneetGopinath/Sanitizers BKS on GitHub
- * @see       https://packagist.org/packages/sanitizers/sanitizers BKS on Packagist
+ * @see       https://github.com/PuneetGopinath/Sanitizers BKS - GitHub
+ * @see       https://packagist.org/packages/sanitizers/sanitizers BKS - Packagist
  */
 
 namespace Sanitizers\Sanitizers;
@@ -81,9 +81,9 @@ class Sanitizer
     /**
      * Sanitizer class constructor.
      *
-     * @param  bool|null                     $exceptions Do you want to enable exceptions?
-     * @param  \Psr\Log\LoggerInterface|null $logger     You can pass an instance of a PSR-3 compatible logger here
-     * @param   SanitizerData|null $sanitizerData The SanitizerData class
+     * @param bool|null                     $exceptions Do you want to enable exceptions?
+     * @param \Psr\Log\LoggerInterface|null $logger     You can pass an instance of a PSR-3 compatible logger here
+     * @param SanitizerData|null $sanitizerData The SanitizerData class
      * @return Sanitizer The Sanitizer class
      */
     public function __construct($exceptions = null, $logger = null, $sanitizerData = null)
@@ -106,7 +106,7 @@ class Sanitizer
      * Warning message
      *
      * @param string      $msg         The Warning message
-     * @param string|null $depreciated If the warning is for depreciated mention it here
+     * @param string|null $depreciated If the warning is for something depreciated, mention it here
      * @param string $deprecatedType What is deprecated, function?, argument?
      * @return string The full Warning message
      */
@@ -114,7 +114,8 @@ class Sanitizer
     {
         $message = "Warning: BK Sanitizers: ";
         if (!empty($depreciated)) {
-            $message .= "Using depreciated " . $deprecatedType . $depreciated . ", " . $msg;
+            $message .= "Using depreciated " . $deprecatedType . $depreciated;
+            $message .= ", " . $msg;
             error_log($message);
             return $message;
         }
@@ -126,7 +127,7 @@ class Sanitizer
     /**
      * Fatal Error message
      *
-     * @param  $msg The Fatal Error message
+     * @param $msg The Fatal Error message
      * @return string The full Fatal Error message
      */
     private function fatal($msg)
@@ -149,7 +150,7 @@ class Sanitizer
     /**
      * Load configuration from ini file
      *
-     * @param  string $file Path to config.ini file
+     * @param string $file Path to config.ini file
      * @return null Nothing to return, So null
      */
     public function configFromIni($file = "config.ini")
@@ -163,8 +164,8 @@ class Sanitizer
     /**
      * Modify configuration options temporarily
      *
-     * @param  string       $case  The key of the setting
-     * @param  string|array $value The value of the setting
+     * @param string       $case  The key of the setting
+     * @param string|array $value The value of the setting
      * @return bool Whether the config option has been modified or not
      */
     public function set($case, $value = "default")
@@ -194,7 +195,9 @@ class Sanitizer
                 break;
             default:
                 if ($this->exceptions) {
-                    throw new \Exception($this->fatal("Invalid config key: $case with value: $value"));
+                    throw new \Exception($this->fatal(
+                        "Invalid config key: $case with value: $value"
+                    ));
                 }
                 return false;
                 break;
@@ -208,7 +211,10 @@ class Sanitizer
                 $this->config["preventXSS"] === true &&
                 strtoupper($this->config["encoding"]) !== "UTF-8"
             ) {
-                $msg = $this->fatal("If you set preventXSS as true then you should also set encoding to \"UTF-8\"");
+                $msg = $this->fatal(
+                    "If you set preventXSS as true" .
+                    " then you should also set encoding to \"UTF-8\""
+                );
                 if ($this->logger) {
                     $this->logger->error($msg);
                 }
@@ -226,16 +232,16 @@ class Sanitizer
     /**
      * Default Sanitizing input function
      *
-     * @param  string $text             The input data.
-     * @param  bool   $trim             Do you want to trim the input data?
-     * @param  bool   $htmlspecialchars Do you want to use htmlspecialchars in input data?
-     * @param  bool   $alpha_num        Is the input data alpha numeric?
-     * @param  bool   $ucwords          Do you want to convert first letter of each word to upper case?
+     * @param string $text             The input data.
+     * @param bool   $trim             Do you want to trim the input data?
+     * @param bool   $htmlspecialchars Do you want to use htmlspecialchars in input data?
+     * @param bool   $alpha_num        Is the input data alpha numeric?
+     * @param bool   $ucwords          Do you want to convert first letter of each word to upper case?
      * @return string The sanitized string
      */
     public function clean($text, $trim = true, $htmlspecialchars = true, $alpha_num = false, $ucwords = false)
     {
-        $text = $this->stripTagsContent((string)$text);
+        $text = strip_tags($this->stripTagsContent((string)$text));
 
         if (
             $htmlspecialchars &&
@@ -246,21 +252,15 @@ class Sanitizer
             $htmlspecialchars &&
             !($this->config["preventXSS"])
         ) {
-            $text = htmlspecialchars($text, ENT_QUOTES, $this->config["encoding"]);
-        }
-
-        if ($trim) {
-            $text = trim($text);
+            $text = htmlspecialchars(
+                $text,
+                ENT_QUOTES,
+                $this->config["encoding"]
+            );
         }
 
         if ($alpha_num) {
             $text = preg_replace("/\W/si", "", $text);
-        }
-
-        $text = str_replace(chr(0)/*NULL character*/, "", $text); //Remove NULL character
-
-        if (mb_strlen($text) > $this->config["maxInputLength"]) {
-            $text = mb_substr($text, 0, $this->config["maxInputLength"]);
         }
 
         if ($this->config["preventXSS"]) {
@@ -268,9 +268,21 @@ class Sanitizer
         }
 
         if (function_exists("iconv") && $this->config["preventXSS"]) {
-            $text = iconv(mb_detect_encoding($text, mb_detect_order(), true), "UTF-8", $text);
+            $text = iconv(
+                mb_detect_encoding($text, mb_detect_order(), true),
+                "UTF-8",
+                $text
+            );
         } elseif (!function_exists("iconv")) {
             $this->warn("PHP extension iconv not installed.");
+        }
+
+        if ($trim) {
+            $text = trim($text);
+        }
+
+        if (mb_strlen($text) > $this->config["maxInputLength"]) {
+            $text = mb_substr($text, 0, $this->config["maxInputLength"]);
         }
 
         if ($ucwords) {
@@ -283,12 +295,12 @@ class Sanitizer
     /**
      * Sanitize a input
      *
-     * @param  string           $type             The input type.
-     * @param  string|int|float $text             The input data.
-     * @param  bool             $trim             Do you want to trim the input data?
-     * @param  bool             $htmlspecialchars Do you want to use php htmlspecialchars function on the input data?
-     * @param  bool             $alpha_num        Is the input data alpha numeric?
-     * @param  bool             $ucwords          Do you want to convert first letter of each word to upper case?
+     * @param string           $type             The input type.
+     * @param string|int|float $text             The input data.
+     * @param bool             $trim             Do you want to trim the input data?
+     * @param bool             $htmlspecialchars Do you want to use php htmlspecialchars function on the input data?
+     * @param bool             $alpha_num        Is the input data alpha numeric?
+     * @param bool             $ucwords          Do you want to convert first letter of each word to upper case?
      * @return string|int|float The sanitized string or int or float
      */
     public function sanitize($type, $text, $trim = true, $htmlspecialchars = true, $alpha_num = false, $ucwords = false)
@@ -297,7 +309,7 @@ class Sanitizer
         if (empty($type)) {
             $type = gettype($text);
 
-            if ($type === "string") {//If $type is not given and php detects it as a string
+            if ($type === "string") {//If $type is not given, php detects it as a string
                 $type = ""; //Then use the default clean function
             }
         }
@@ -307,16 +319,37 @@ class Sanitizer
         switch (strtolower($type)) {
             case "int":
             case "integer":
-                $text = preg_replace("/[^0-9]/s", "", filter_var($text, FILTER_SANITIZE_NUMBER_INT));
+                $text = preg_replace(
+                    "/[^0-9]/s",
+                    "",
+                    filter_var(
+                        $text,
+                        FILTER_SANITIZE_NUMBER_INT
+                    )
+                );
                 $text = (int)intval($text) + 0;
                 break;
             case "float":
-                $text = filter_var($text, FILTER_SANITIZE_NUMBER_FLOAT, FILTER_FLAG_ALLOW_FRACTION);
-                $text = (float)floatval(preg_replace("/[^0-9.]/s", "", $text)) + 0;
+                $text = preg_replace(
+                    "/[^0-9.]/s",
+                    "",
+                    filter_var(
+                        $text,
+                        FILTER_SANITIZE_NUMBER_FLOAT,
+                        FILTER_FLAG_ALLOW_FRACTION
+                    )
+                );
+                $text = (float)floatval($text) + 0;
                 break;
             case "string":
             case "text":
-                $text = $this->clean($text, $trim, $htmlspecialchars, $alpha_num, $ucwords);
+                $text = $this->clean(
+                    $text,
+                    $trim,
+                    $htmlspecialchars,
+                    $alpha_num,
+                    $ucwords
+                );
                 $text = filter_var(
                     $text,
                     FILTER_SANITIZE_STRING,
@@ -331,10 +364,19 @@ class Sanitizer
                     FILTER_FLAG_STRIP_LOW | FILTER_FLAG_STRIP_HIGH
                 );
                 $text = preg_replace("/[^a-f0-9]/s", "", $text);
-                $text = $this->clean($text, $trim, $htmlspecialchars, true, false);
+                $text = $this->clean(
+                    $text,
+                    $trim,
+                    $htmlspecialchars,
+                    true,
+                    false
+                );
                 break;
             case "url":
-                $text = filter_var($this->stripTagsContent($text), FILTER_SANITIZE_URL);
+                $text = filter_var(
+                    $this->stripTagsContent($text),
+                    FILTER_SANITIZE_URL
+                );
                 break;
             case "password":
                 $text = filter_var(
@@ -351,21 +393,40 @@ class Sanitizer
                     FILTER_FLAG_STRIP_LOW | FILTER_FLAG_STRIP_HIGH
                 );
                 $text = preg_replace("/[^A-Za-z\s+]/s", "", $text);
-                $text = $this->clean($text, $trim, $htmlspecialchars, $alpha_num, true);
+                $text = $this->clean(
+                    $text,
+                    $trim,
+                    $htmlspecialchars,
+                    $alpha_num,
+                    true
+                );
                 break;
             case "message":
                 $text = $this->clean($text, false, true, false, false);
                 break;
             case "email":
-                $text = $this->clean($text, $trim, $htmlspecialchars, false, false);
                 $text = filter_var(strtolower($text), FILTER_SANITIZE_EMAIL);
+                $text = $this->clean($text, $trim, $htmlspecialchars, false, false);
                 break;
             case "username":
-                $text = $this->sanitize("text", $text, $trim, $htmlspecialchars, true, false);
+                $text = $this->sanitize(
+                    "text",
+                    $text,
+                    $trim,
+                    $htmlspecialchars,
+                    true,
+                    false
+                );
                 $text = preg_replace("/[^a-z0-9]/s", "", strtolower($text));
                 break;
             default:
-                $text = $this->clean($text, $trim, $htmlspecialchars, $alpha_num, $ucwords);
+                $text = $this->clean(
+                    $text,
+                    $trim,
+                    $htmlspecialchars,
+                    $alpha_num,
+                    $ucwords
+                );
                 break;
         }
         if ($this->logger) {
@@ -381,19 +442,27 @@ class Sanitizer
     /**
      * Text without any number (Depreciated)
      *
-     * @param  string $text             The input data.
-     * @param  bool   $trim             Do you want to trim the input data?
-     * @param  bool   $htmlspecialchars Do you want to use php htmlspecialchars function on the input data?
-     * @param  bool   $alpha_num
+     * @param string $text             The input data.
+     * @param bool   $trim             Do you want to trim the input data?
+     * @param bool   $htmlspecialchars Do you want to use php htmlspecialchars function on the input data?
+     * @param bool   $alpha_num
      * @return string The sanitized string
      */
     public function NonNumericText($text, $trim = true, $htmlspecialchars = true, $alpha_num = false, $ucwords = false)
     {
         $this->warn(
-            "use sanitize function with type param as \"text\" (see functions in docs for more info)",
+            "use sanitize function with type param as \"text\"" .
+            "(see functions in docs for more info)",
             "NonNumericText"
         );
-        $text = $this->sanitize("text", (string)$text, $trim, $htmlspecialchars, $alpha_num, $ucwords);
+        $text = $this->sanitize(
+            "text",
+            $text,
+            $trim,
+            $htmlspecialchars,
+            $alpha_num,
+            $ucwords
+        );
         $text = preg_replace("/[^A-Za-z]/s", "", $text);
         return $text;
     }
@@ -401,9 +470,9 @@ class Sanitizer
     /**
      * https://www.php.net/manual/en/function.strip-tags.php#86964
      *
-     * @param  string $text   The html code.
-     * @param  string $tags   The allowed html tags.
-     * @param  bool   $invert Do you want to change parameter #2 to `The disallowed html tags.`?
+     * @param string $text   The html code.
+     * @param string $tags   The allowed html tags.
+     * @param bool   $invert Do you want to change param 2 to `The disallowed html tags.`?
      * @return string The striped string
      */
     private function stripTagsContent($text, $tags = "", $invert = false)
@@ -413,12 +482,24 @@ class Sanitizer
 
         if (is_array($tags) && count($tags) > 0) {
             if ($invert === false) {
-                return preg_replace('@<(?!(?:' . implode("|", $tags) . ')\b)(\w+)\b.*?>.*?</\1>@si', "", $text);
+                return preg_replace(
+                    '@<(?!(?:' . implode("|", $tags) . ')\b)(\w+)\b.*?>.*?</\1>@si',
+                    "",
+                    $text
+                );
             } else {
-                return preg_replace('@<(' . implode("|", $tags) . ')\b.*?>.*?</\1>@si', "", $text);
+                return preg_replace(
+                    '@<(' . implode("|", $tags) . ')\b.*?>.*?</\1>@si',
+                    "",
+                    $text
+                );
             }
         } elseif ($invert === false) {
-            return preg_replace('@<(\w+)\b.*?>.*?</\1>@si', "", $text);
+            return preg_replace(
+                '@<(\w+)\b.*?>.*?</\1>@si',
+                "",
+                $text
+            );
         }
         return $text;
     }
@@ -426,21 +507,28 @@ class Sanitizer
     /**
      * Sanitize a html code
      *
-     * @param  string $text
-     * @param  string $tags
+     * @param string $text
+     * @param string $tags
      * @return string The sanitized html code
      */
     public function HTML($text, $tags = "<b><i><em><p><a><br>")
     {
         // Remove any attribute starting with on or xmlns
-        $text = preg_replace('#(<[^>]+?[\x00-\x20"\'])(?:on|xmlns)[^>]*+>#iu', '$1>', (string)$text);
+        $text = preg_replace(
+            '#(<[^>]+?[\x00-\x20"\'])(?:on|xmlns)[^>]*+>#iu',
+            '$1>',
+            (string)$text
+        );
         $text = $this->stripTagsContent((string)$text, $tags);
         if (class_exists("\HTMLPurifier")) {
             $config = \HTMLPurifier_Config::createDefault();
             $purifier = new \HTMLPurifier($config);
             $clean_html = $purifier->purify($text);
         } else {
-            $this->warn("HTMLPurifier is not installed or required (See INSTALL.md for more info).");
+            $this->warn(
+                "HTMLPurifier is not installed or required " .
+                "(See INSTALL.md for more info)."
+            );
             $clean_html = $text;
         }
 
@@ -450,14 +538,19 @@ class Sanitizer
     /**
      * Sanitize a array
      *
-     * @param  array $array The input array.
-     * @param  array $filters The filters to apply to the values in the array
+     * @param array $array The input array.
+     * @param array $filters The filters to apply to the values in the array
      * @return array The sanitized array
      */
     public function sanitizeArray($array, $filters = array("types" => array()))
     {
         foreach ($array as $key => $value) {
-            $settings = array("trim" => true, "htmlspecialchars" => true, "alpha_num" => false, "ucwords" => false);
+            $settings = array(
+                "trim" => true,
+                "htmlspecialchars" => true,
+                "alpha_num" => false,
+                "ucwords" => false
+            );
             if (isset($filters[$key])) {
                 foreach ($settings as $key2 => $value2) {
                     if (isset($filters[$key][$key2])) {
@@ -473,7 +566,10 @@ class Sanitizer
                 }
             }
 
-            if (empty($filters[$key]["tags"]) && $filters["types"][$key] === "html") {
+            if (
+                empty($filters[$key]["tags"]) &&
+                $filters["types"][$key] === "html"
+            ) {
                 $filters[$key]["tags"] = "<b><i><em><p><a><br>";
             }
 
@@ -492,10 +588,24 @@ class Sanitizer
                     break;
                 case "string":
                 case "text":
-                    $sanitized = $this->sanitize("text", $value, $trim, $htmlspecialchars, $alpha_num, $ucwords);
+                    $sanitized = $this->sanitize(
+                        "text",
+                        $value,
+                        $trim,
+                        $htmlspecialchars,
+                        $alpha_num,
+                        $ucwords
+                    );
                     break;
                 case "hex":
-                    $sanitized = $this->sanitize("hex", $value, $trim, $htmlspecialchars, true, false);
+                    $sanitized = $this->sanitize(
+                        "hex",
+                        $value,
+                        $trim,
+                        $htmlspecialchars,
+                        true,
+                        false
+                    );
                     break;
                 case "url":
                     $sanitized = $this->sanitize("url", $value);
@@ -504,22 +614,55 @@ class Sanitizer
                     $sanitized = $this->sanitize("password", $value);
                     break;
                 case "name":
-                    $sanitized = $this->sanitize("name", $value, $trim, $htmlspecialchars, true, $ucwords);
+                    $sanitized = $this->sanitize(
+                        "name",
+                        $value,
+                        $trim,
+                        $htmlspecialchars,
+                        true,
+                        $ucwords
+                    );
                     break;
                 case "message":
-                    $sanitized = $this->clean($value, false, true, false, false);
+                    $sanitized = $this->clean(
+                        $value,
+                        false,
+                        true,
+                        false,
+                        false
+                    );
                     break;
                 case "email":
-                    $sanitized = $this->sanitize("email", $value, $trim, $htmlspecialchars, false, false);
+                    $sanitized = $this->sanitize(
+                        "email",
+                        $value,
+                        $trim,
+                        $htmlspecialchars,
+                        false,
+                        false
+                    );
                     break;
                 case "username":
-                    $sanitized = $this->sanitize("username", $value, $trim, $htmlspecialchars, true, false);
+                    $sanitized = $this->sanitize(
+                        "username",
+                        $value,
+                        $trim,
+                        $htmlspecialchars,
+                        true,
+                        false
+                    );
                     break;
                 case "html":
                     $sanitized = $this->HTML($value, $filters[$key]["tags"]);
                     break;
                 default:
-                    $sanitized = $this->clean($value, $trim, $htmlspecialchars, $alpha_num, $ucwords);
+                    $sanitized = $this->clean(
+                        $value,
+                        $trim,
+                        $htmlspecialchars,
+                        $alpha_num,
+                        $ucwords
+                    );
                     break;
             }